change(document): move algorithm chapter from extra1 to step3 (#255)
## What
This PR moves the algorithm chapter from extra 1 to step 3 as we've
decided moving the algorithm chapter from optional to basic.

I think step 3 is the best place to locate this chapter since the order
that Git -> setting up environment -> algorithm/data structure -> web
API implementations -> ... is natural for students to follow.

Slack discussion:
https://mercari.slack.com/archives/G01HQNGEK8F/p1740703480870359

## CHECKS :warning:

Please make sure you are aware of the following.

- [ ] **The changes in this PR doesn't have private information