change(go): revise Go code (#251)
## What
I've revised Go code as I noticed that it would be easier if we
introduced interface from step 3.

Also, this PR removes Echo framework as we can replace it with standard
library, net/http and log/slog.
I think Echo framework is easy to write, but it's hard for the beginner
to get a big picture.
I think it's worth for a beginner to understand whole code.

## CHECKS :warning:

Please make sure you are aware of the following.

- [ ] **The changes in this PR doesn't have private information